+++ src/cmd/app/main.go
@@ -10,7 +10,7 @@ import (
 
 func init() {
 	env.Vars.SystemName = "box"
-	env.Version = "24.0.9"
+	env.Version = "24.0.10"
 
 	if env.Mode == env_mode.Dev {
 		if err := tracer.Init(); err != nil {