+++ src/cmd/scripts/gqlgen.go
@@ -1,0 +1,10 @@
+//go:build ignore
+// +build ignore
+
+package main
+
+import "github.com/99designs/gqlgen/graphql/cmd"
+
+func main() {
+	cmd.Execute()
+}