+++ src/internal/app/transports/rest_api/config.go
@@ -1,0 +1,4 @@
+package rest_api
+
+type Config struct {
+}