+++ src/internal/app/transports/rest_api/engine.go
@@ -1,6 +1,9 @@
-package graphql
+package rest_api
 
-import "context"
+import (
+	"context"
+	"github.com/gofiber/fiber/v3"
+)
 
 type Engine interface {
 	Serve() (err error)
@@ -9,6 +12,8 @@ type Engine interface {
 
 func New(ctx context.Context) (eng Engine, err error) {
 	var e = new(engine)
+
+	e.app = fiber.New()
 
 	eng = e