+++ src/internal/app/transports/rest_api/router.go
@@ -1,0 +1,6 @@
+package rest_api
+
+func (e *engine) initRouter() (err error) {
+
+	return
+}