+++ src/internal/entities/role/entity.go
@@ -1,0 +1,13 @@
+package role
+
+type Role struct {
+	ID        int64
+	ProjectID int64
+
+	Title        string
+	Inheritances []*Inheritance
+}
+
+type Inheritance struct {
+	*Role
+}