+++ src/pkg/core/addons/encryption_keys/errors.go
@@ -1,0 +1,7 @@
+package encryption_keys
+
+import "errors"
+
+var (
+	ErrInvalidEncryptionKeys = errors.New("Invalid encryption keys. ")
+)