+++ src/pkg/core/addons/encryption_keys/tasks.go
@@ -1,0 +1,21 @@
+package encryption_keys
+
+import (
+	"context"
+	"sm-box/src/pkg/core/tools/task_scheduler"
+)
+
+var (
+	TaskInitEncryptionKeys = task_scheduler.Task{
+		Name: "Init encryption keys",
+		Type: task_scheduler.TaskBeforeBoot,
+		Func: func(ctx context.Context) (err error) {
+
+			if err = Init(); err != nil {
+				return
+			}
+
+			return
+		},
+	}
+)