+++ src/pkg/core/addons/pid/errors.go
@@ -1,0 +1,8 @@
+package pid
+
+import "errors"
+
+var (
+	ErrPidFileAlreadyExist = errors.New("PID file already exist. ")
+	ErrPidFileNotExist     = errors.New("PID file not exist. ")
+)