+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # ChangeLog:
 
+### v24.0.10:
+- Доработка архитектуры проекта;
+- Доработки основы [коробки](src/internal/app/box.go);
+- Разработка [дополнения ядра для использования ключей шифрования в работе системы](src/pkg/core/addons/encryption_keys/encryption_keys.go);
+- Разработка [дополнения ядра для записи PID файла системы](src/pkg/core/addons/pid/pid.go);
+- Доработки [переменных окружения](src/pkg/core/env/vars/vars.go);
+
+---
+
 ### v24.0.9:
 - Переработка архитектуры проекта;
 - Доработки основы [коробки](src/internal/app/box.go);
@@ -24,7 +33,7 @@
 ### v24.0.6:
 - Доработки компонента [ведения журнала трессировки вызовов](src/pkg/core/components/tracer/tracer.go);
 - Разработан [инструмент для управления задачами ядра системы](src/pkg/core/tools/task_scheduler/task_scheduler.go);
-- Компонент для [компонент для корректного завершения работы ядра системы](src/pkg/core/components/closer/closer.go) стал [инструментом](src/pkg/core/tools/closer/closer.go);
+- Компонент для [компонент для корректного завершения работы ядра системы](src/pkg/core/tools/closer/closer.go) стал [инструментом](src/pkg/core/tools/closer/closer.go);
 - Добавлено взаимодействие с [инструментами ядра](src/pkg/core/tools) в [ядре системы](src/pkg/core/core.go);
 - Добавлен уровень ведения журнала трессировки для инструментов ядра в компоненте [ведения журнала трессировки вызовов](src/pkg/core/components/tracer/tracer.go);